+++ tests/testthat/_ref/my-analysis.R
@@ -1,4 +1,5 @@
-# Read my streets data ---------------
+# Analyse my streets ---------------
+## Read my streets data ---------------
 my_streets <- read.csv("data/my-streets.csv")
 new_dat <- my_streets |>
   dplyr::mutate(